refactor: extracted `VideoRendererProgressPanel`
