+++ example/lib/features/render/video_renderer_page.dart
@@ -10,6 +10,7 @@ import 'package:media_kit_video/media_kit_video.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:pro_video_editor/pro_video_editor.dart';
 import 'package:pro_video_editor_example/shared/utils/render_cancel_capability.dart';
+import 'package:pro_video_editor_example/shared/widgets/video_renderer_progress.dart';
 
 import '/core/constants/example_filters.dart';
 import '/shared/utils/bytes_formatter.dart';
@@ -525,60 +526,6 @@ class _VideoRendererPageState extends State<VideoRendererPage> {
           subtitle: const Text('35 Mbps bitrate'),
         ),
       ],
-    );
-  }
-}
-
-/// Progress indicator panel displayed while the renderer is exporting a video.
-class VideoRendererProgressPanel extends StatelessWidget {
-  /// Creates a [VideoRendererProgressPanel].
-  const VideoRendererProgressPanel({
-    super.key,
-    required this.progressStream,
-    required this.supportsCancel,
-    this.onCancel,
-  });
-
-  /// Emits [ProgressModel] updates for the active render task.
-  final Stream<ProgressModel> progressStream;
-
-  /// Whether the current platform exposes a cancel API.
-  final bool supportsCancel;
-
-  /// Invoked when the cancel button is tapped.
-  final VoidCallback? onCancel;
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<ProgressModel>(
-      stream: progressStream,
-      builder: (context, snapshot) {
-        final double progress = snapshot.data?.progress ?? 0;
-
-        return TweenAnimationBuilder<double>(
-          tween: Tween<double>(begin: 0, end: progress),
-          duration: const Duration(milliseconds: 300),
-          builder: (context, animatedValue, _) {
-            return Column(
-              spacing: 12,
-              children: [
-                CircularProgressIndicator(
-                  value: animatedValue,
-                  // ignore: deprecated_member_use
-                  year2023: false,
-                ),
-                Text('${(animatedValue * 100).toStringAsFixed(1)} / 100'),
-                if (supportsCancel && onCancel != null)
-                  FilledButton.icon(
-                    onPressed: onCancel,
-                    icon: const Icon(Icons.stop_circle_outlined),
-                    label: const Text('Cancel render'),
-                  ),
-              ],
-            );
-          },
-        );
-      },
     );
   }
 }
